+++ server/src/services/utilities.ts
@@ -1228,7 +1228,7 @@ export function isTagName(node: Node): boolean {
 
 /** @internal */
 export function getContextualTypeFromParentOrAncestorTypeNode(node: Expression, checker: TypeChecker): Type | undefined {
-    if (node.flags & (NodeFlags.JSDoc & ~NodeFlags.JavaScriptFile)) return undefined;
+    if (node.flags & (NodeFlags.JSDoc /*& ~NodeFlags.JavaScriptFile*/)) return undefined;
 
     const contextualType = getContextualTypeFromParent(node, checker);
     if (contextualType) return contextualType;