+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.21
 
 -   Fix: [Foreach over an array always results in a string type #161](https://github.com/jlchmura/lpc-language-server/issues/161)
+-   Fix: [Variable in foreach losing type #163](https://github.com/jlchmura/lpc-language-server/issues/163)
 -   Fix: [Flow type lost inside inline closure #164](https://github.com/jlchmura/lpc-language-server/issues/164)
 -   Improved efun typing:
     -   FluffOS `unique_array`